+++ artifact.go
@@ -9,12 +9,12 @@ import (
 	"path/filepath"
 
 	"github.com/gin-gonic/gin"
-	"github.com/iver-wharf/wharf-api/internal/ctxparser"
-	"github.com/iver-wharf/wharf-api/internal/wherefields"
-	"github.com/iver-wharf/wharf-api/pkg/model/database"
-	"github.com/iver-wharf/wharf-api/pkg/model/response"
-	"github.com/iver-wharf/wharf-api/pkg/modelconv"
-	"github.com/iver-wharf/wharf-api/pkg/orderby"
+	"github.com/iver-wharf/wharf-api/v5/internal/ctxparser"
+	"github.com/iver-wharf/wharf-api/v5/internal/wherefields"
+	"github.com/iver-wharf/wharf-api/v5/pkg/model/database"
+	"github.com/iver-wharf/wharf-api/v5/pkg/model/response"
+	"github.com/iver-wharf/wharf-api/v5/pkg/modelconv"
+	"github.com/iver-wharf/wharf-api/v5/pkg/orderby"
 	"github.com/iver-wharf/wharf-core/pkg/ginutil"
 	"gorm.io/gorm"
 )