+++ pkg/orderby/column.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/iver-wharf/wharf-api/pkg/model/database"
+	"github.com/iver-wharf/wharf-api/v5/pkg/model/database"
 	"gorm.io/gorm/clause"
 )