+++ pkg/orderby/orderby_test.go
@@ -3,7 +3,7 @@ package orderby
 import (
 	"testing"
 
-	"github.com/iver-wharf/wharf-api/pkg/model/database"
+	"github.com/iver-wharf/wharf-api/v5/pkg/model/database"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )